drivers: format the code
Format the code.
