+++ drivers/audio/Kconfig.bee
@@ -1,13 +1,11 @@
-# BEE codec configuration
-
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright (c) 2026 Realtek Semiconductor Corp.
 # SPDX-License-Identifier: Apache-2.0
 
 config AUDIO_BEE
-	bool "BEE MCU codec controller driver"
+	bool "CODEC controller driver for Realtek Bee series SoC"
 	default y
 	depends on DT_HAS_REALTEK_BEE_CODEC_ENABLED
 	select I2S
 	select USE_HAL_REALTEK_CODEC
 	help
-	  Enable the bee codec driver.
+	  This option enables the CODEC driver for Realtek Bee SoCs.
