+++ drivers/can/can_bee.c
@@ -31,10 +31,6 @@
 #endif
 #endif
 
-#include <trace.h>
-
-#define DBG_DIRECT_SHOW 0
-
 LOG_MODULE_REGISTER(can_bee, CONFIG_CAN_LOG_LEVEL);
 struct can_bee_mb_data {
 	uint8_t idx;
@@ -99,11 +95,10 @@ static int can_bee_check_bus_on(k_timeout_t timeout)
 
 static void can_bee_tx_complete(const struct device *dev, int mb_list_idx)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	struct can_bee_data *data = dev->data;
 	can_tx_callback_t callback = data->tx_mb_list[mb_list_idx].tx_callback;
+
+	LOG_DBG("[%s]", __func__);
 
 	if (data->tx_mb_list[mb_list_idx].is_busy) {
 		data->tx_mb_list[mb_list_idx].is_busy = false;
@@ -119,29 +114,26 @@ static void can_bee_tx_complete(const struct device *dev, int mb_list_idx)
 
 static void can_bee_rx_complete(const struct device *dev, int mb_list_idx)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	struct can_bee_data *data = dev->data;
 	can_rx_callback_t callback = NULL;
 	CANMsgBufInfo_TypeDef mb_info;
 	CANDataFrameSel_TypeDef frame_type;
 	struct can_frame frame;
 
+	LOG_DBG("[%s]", __func__);
+
 	CAN_GetMsgBufInfo(data->rx_mb_list[mb_list_idx].idx, &mb_info);
 	memset(frame.data, 0, 8);
 	CAN_GetRamData(mb_info.data_length, frame.data);
 	frame_type = CAN_CheckFrameType(mb_info.rtr_bit, mb_info.ide_bit);
 
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[CAN HANDLER] rx_mb_list[mb_list_idx].idx%d frame_type %d, frame_id = 0x%03x, "
+	LOG_DBG("[CAN HANDLER] rx_mb_list[mb_list_idx].idx%d frame_type %d, frame_id = 0x%03x, "
 		   "ext_frame_id = 0x%05x",
 		   data->rx_mb_list[mb_list_idx].idx, frame_type, mb_info.standard_frame_id,
 		   mb_info.extend_frame_id);
 	for (uint8_t index = 0; index < mb_info.data_length; index++) {
-		DBG_DIRECT("[CAN HANDLER] frame.data [%d] 0x%02x", index, frame.data[index]);
+		LOG_DBG("[CAN HANDLER] frame.data [%d] 0x%02x", index, frame.data[index]);
 	}
-#endif
 
 	frame.dlc = mb_info.data_length;
 	frame.flags = (mb_info.rtr_bit ? CAN_FRAME_RTR : 0) | (mb_info.ide_bit ? CAN_FRAME_IDE : 0);
@@ -156,11 +148,10 @@ static void can_bee_rx_complete(const struct device *dev, int mb_list_idx)
 
 static int can_bee_get_capabilities(const struct device *dev, can_mode_t *cap)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	ARG_UNUSED(dev);
 	ARG_UNUSED(cap);
+
+	LOG_DBG("[%s]", __func__);
 
 	*cap = CAN_MODE_NORMAL | CAN_MODE_LISTENONLY | CAN_MODE_MANUAL_RECOVERY;
 	return 0;
@@ -168,13 +159,12 @@ static int can_bee_get_capabilities(const struct device *dev, can_mode_t *cap)
 
 static int can_bee_start(const struct device *dev)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s] dev %s", dev->name, __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	int ret = 0;
 	CAN_InitTypeDef *init_struct = &(data->init_struct);
+
+	LOG_DBG("[%s] dev %s", dev->name, __func__);
 
 	k_mutex_lock(&data->inst_mutex, K_FOREVER);
 
@@ -217,12 +207,11 @@ unlock:
 
 static int can_bee_stop(const struct device *dev)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s] dev %s", dev->name, __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	int ret = 0;
+
+	LOG_DBG("[%s] dev %s", dev->name, __func__);
 
 	k_mutex_lock(&data->inst_mutex, K_FOREVER);
 
@@ -255,13 +244,10 @@ unlock:
 
 static int can_bee_set_mode(const struct device *dev, can_mode_t mode)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s] dev %s mode 0x%x", __func__, dev->name, mode);
-#endif
 	struct can_bee_data *data = dev->data;
 	CAN_InitTypeDef *init_struct = &(data->init_struct);
 
-	LOG_DBG("Set mode %d", mode);
+	LOG_DBG("[%s] dev %s mode 0x%x", __func__, dev->name, mode);
 
 	if ((mode & ~(CAN_MODE_LOOPBACK | CAN_MODE_LISTENONLY | CAN_MODE_ONE_SHOT)) != 0) {
 		LOG_ERR("unsupported mode: 0x%08x", mode);
@@ -297,11 +283,10 @@ static int can_bee_set_mode(const struct device *dev, can_mode_t mode)
 
 static int can_bee_set_timing(const struct device *dev, const struct can_timing *timing)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s] dev %s", __func__, dev->name);
-#endif
 	struct can_bee_data *data = dev->data;
 	CAN_InitTypeDef *init_struct = &(data->init_struct);
+
+	LOG_DBG("[%s] dev %s", __func__, dev->name);
 
 	if (data->common.started) {
 		return -EBUSY;
@@ -337,9 +322,6 @@ static int can_bee_request_message_buffer(const struct device *dev, bool is_tx)
 static int can_bee_send(const struct device *dev, const struct can_frame *frame,
 			k_timeout_t timeout, can_tx_callback_t callback, void *user_data)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	CAN_TypeDef *can = cfg->can;
@@ -445,16 +427,15 @@ fail:
 static int can_bee_add_rx_filter(const struct device *dev, can_rx_callback_t callback,
 				 void *user_data, const struct can_filter *filter)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s] filter->id=0x%x, filter->flags=0x%x, filter->mask=0x%x", __func__,
-		   filter->id, filter->flags, filter->mask);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	CAN_TypeDef *can = cfg->can;
 	CANRxFrame_TypeDef rx_frame_type;
 	CANError_TypeDef rx_error;
 	int mb_list_idx;
+
+	LOG_DBG("[%s] filter->id=0x%x, filter->flags=0x%x, filter->mask=0x%x", __func__,
+		   filter->id, filter->flags, filter->mask);
 
 	if ((filter->flags & ~(CAN_FILTER_IDE)) != 0) {
 		LOG_ERR("Unsupported CAN filter flags 0x%02x", filter->flags);
@@ -523,13 +504,12 @@ unlock:
 
 static void can_bee_remove_rx_filter(const struct device *dev, int mb_list_idx)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	CAN_TypeDef *can = cfg->can;
 	CANRxFrame_TypeDef rx_frame_type;
+
+	LOG_DBG("[%s]", __func__);
 
 	k_mutex_lock(&filter_mutex, K_FOREVER);
 	k_mutex_lock(&data->inst_mutex, K_FOREVER);
@@ -551,12 +531,11 @@ static void can_bee_remove_rx_filter(const struct device *dev, int mb_list_idx)
 static int can_bee_get_state(const struct device *dev, enum can_state *state,
 			     struct can_bus_err_cnt *err_cnt)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	CAN_TypeDef *can = cfg->can;
+
+	LOG_DBG("[%s]", __func__);
 
 	if (state != NULL) {
 		if (!data->common.started) {
@@ -582,13 +561,12 @@ static int can_bee_get_state(const struct device *dev, enum can_state *state,
 #if defined(CONFIG_CAN_MANUAL_RECOVERY_MODE)
 static int can_bee_recover(const struct device *dev, k_timeout_t timeout)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	CAN_TypeDef *can = cfg->can;
 	int ret = -EAGAIN;
+
+	LOG_DBG("[%s]", __func__);
 
 	if (!data->common.started) {
 		return -ENETDOWN;
@@ -612,10 +590,9 @@ static int can_bee_recover(const struct device *dev, k_timeout_t timeout)
 static void can_bee_set_state_change_callback(const struct device *dev,
 					      can_state_change_callback_t cb, void *user_data)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	struct can_bee_data *data = dev->data;
+
+	LOG_DBG("[%s]", __func__);
 
 	data->common.state_change_cb = cb;
 	data->common.state_change_cb_user_data = user_data;
@@ -629,10 +606,9 @@ static void can_bee_set_state_change_callback(const struct device *dev,
 
 static int can_bee_get_core_clock(const struct device *dev, uint32_t *rate)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	ARG_UNUSED(dev);
+
+	LOG_DBG("[%s]", __func__);
 
 	*rate = 40000000;
 	return 0;
@@ -640,11 +616,10 @@ static int can_bee_get_core_clock(const struct device *dev, uint32_t *rate)
 
 static int can_bee_get_max_filters(const struct device *dev, bool ide)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	ARG_UNUSED(dev);
 	ARG_UNUSED(ide);
+
+	LOG_DBG("[%s]", __func__);
 
 	if (ide) {
 		return CONFIG_CAN_BEE_MAX_EXT_ID_FILTER;
@@ -655,14 +630,13 @@ static int can_bee_get_max_filters(const struct device *dev, bool ide)
 
 static inline void can_bee_isr(const struct device *dev)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	struct can_bee_data *data = dev->data;
 	struct can_bus_err_cnt err_cnt;
 	enum can_state state = 0;
 	const can_state_change_callback_t cb = data->common.state_change_cb;
 	void *state_change_cb_data = data->common.state_change_cb_user_data;
+
+	LOG_DBG("[%s]", __func__);
 
 	if (SET == CAN_GetINTStatus(CAN_BUS_OFF_INT_FLAG)) {
 		CAN_ClearINTPendingBit(CAN_BUS_OFF_INT_FLAG);
@@ -876,14 +850,13 @@ static const struct can_driver_api can_bee_driver_api = {
 
 static int can_bee_init(const struct device *dev)
 {
-#if DBG_DIRECT_SHOW
-	DBG_DIRECT("[%s]", __func__);
-#endif
 	const struct can_bee_config *cfg = dev->config;
 	struct can_bee_data *data = dev->data;
 	struct can_timing timing = {0};
 	CAN_InitTypeDef *init_struct = &(data->init_struct);
 	int ret;
+
+	LOG_DBG("[%s]", __func__);
 
 	k_mutex_init(&filter_mutex);
 	k_mutex_init(&data->inst_mutex);
