+++ drivers/clock_control/Kconfig.bee
@@ -1,12 +1,10 @@
-# BEE clock control configuration
-
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright (c) 2026 Realtek Semiconductor Corp.
 # SPDX-License-Identifier: Apache-2.0
 
 config CLOCK_CONTROL_BEE
-	bool "Bee clock control"
+	bool "clock controller driver for Realtek Bee series SoC"
 	default y
 	depends on DT_HAS_REALTEK_BEE_CCTL_ENABLED
 	select USE_HAL_REALTEK_RCC
 	help
-	  Enable driver for Bee Reset Clock Unit (RCU).
+	  This option enables the clock controller driver for Realtek Bee SoCs.
