+++ drivers/pwm/Kconfig.bee
@@ -1,13 +1,11 @@
-# BEE pwm configuration
-
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright (c) 2026 Realtek Semiconductor Corp.
 # SPDX-License-Identifier: Apache-2.0
 
 config PWM_BEE
-	bool "BEE PWM driver"
+	bool "PWM driver for Realtek Bee series SoC"
 	default y
 	depends on DT_HAS_REALTEK_BEE_PWM_ENABLED
 	select USE_HAL_REALTEK_TIM
 	select USE_HAL_REALTEK_ENHTIM
 	help
-	  Enable the bee pwm driver.
+	  This option enables the PWM driver for Realtek Bee SoCs.
