+++ drivers/usb/udc/Kconfig.bee
@@ -1,11 +1,11 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright (c) 2026 Realtek Semiconductor Corp.
 # SPDX-License-Identifier: Apache-2.0
 
 config UDC_BEE
-	bool "BEE MCU udc driver"
+	bool "UDC driver for Realtek Bee series SoC"
 	default y
 	depends on DT_HAS_REALTEK_BEE_UDC_ENABLED
 	select USE_HAL_REALTEK_USB
 
 	help
-	  Enables the bee usb driver.
+	  This option enables the UDC driver for Realtek Bee SoCs.
