+++ dts/bindings/can/realtek,bee-can.yaml
@@ -1,4 +1,8 @@
-description: BEE CAN controller
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
+# SPDX-License-Identifier: Apache-2.0
+
+description: |
+  Bee CAN controller.
 
 compatible: "realtek,bee-can"
