+++ dts/bindings/dma/realtek,bee-dma-base.yaml
@@ -1,4 +1,4 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 include: dma-controller.yaml
