+++ dts/bindings/gpio/realtek,bee-gpio.yaml
@@ -1,8 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  BEE GPIO.
+  Bee GPIO.
   Note that some of different pins could be mapped to the same gpio number.
 
   For RTL8762GTU and RTL8762GTH,
