+++ dts/bindings/i2c/realtek,bee-i2c.yaml
@@ -1,8 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  BEE I2C.
+  Bee I2C.
 
 compatible: "realtek,bee-i2c"
