+++ dts/bindings/i2s/realtek,bee-i2s.yaml
@@ -1,8 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  BEE I2S.
+  Bee I2S.
 
 compatible: "realtek,bee-i2s"
