+++ dts/bindings/kscan/realtek,bee-kscan.yaml
@@ -1,8 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  BEE KSCAN.
+  Bee KSCAN.
 
   The BEE KSCAN should be used with an external kscan matrix.
   There is a total of 12 rows and 20 columns.
