+++ dts/bindings/sdhc/realtek,bee-sdhc.yaml
@@ -1,8 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
 description: |
-  BEE SDHC controller.
+  Bee SDHC controller.
 
 compatible: "realtek,bee-sdhc"
