+++ dts/bindings/spi/realtek,bee-spi.yaml
@@ -1,7 +1,8 @@
-# Copyright(c) 2025, Realtek Semiconductor Corporation.
+# Copyright(c) 2026, Realtek Semiconductor Corporation.
 # SPDX-License-Identifier: Apache-2.0
 
-description: BEE SPI
+description: |
+   Bee SPI
 
 compatible: "realtek,bee-spi"
