+++ soc/realtek/bee/rtl87x2g/pinctrl_soc.h
@@ -1,16 +1,11 @@
 /*
- * Copyright (c) 2025 Realtek Semiconductor Corp.
+ * Copyright (c) 2026 Realtek Semiconductor Corp.
  *
  * SPDX-License-Identifier: Apache-2.0
  */
 
-/**
- * @file
- * Bee SoC specific helpers for pinctrl driver
- */
-
-#ifndef ZEPHYR_SOC_ARM_REALTEK_RTL_COMMON_PINCTRL_SOC_H_
-#define ZEPHYR_SOC_ARM_REALTEK_RTL_COMMON_PINCTRL_SOC_H_
+#ifndef ZEPHYR_SOC_REALTEK_BEE_RTL87X2G_PINCTRL_SOC_H_
+#define ZEPHYR_SOC_REALTEK_BEE_RTL87X2G_PINCTRL_SOC_H_
 
 #include <zephyr/devicetree.h>
 #include <zephyr/dt-bindings/pinctrl/rtl87x2g-pinctrl.h>
@@ -20,9 +15,17 @@
 extern "C" {
 #endif
 
-/** @cond INTERNAL_HIDDEN */
-
 typedef struct {
+	/* bit[0:10]   pad number
+	 * bit[11:12]  pad pull level
+	 * bit[13]     pad output level
+	 * bit[14]     pad direction
+	 * bit[15]     pad pull strength
+	 * bit[16:31]  pad pinmux function
+	 * bit[32]     pad wakeup high level
+	 * bit[33]     pad wakeup low level
+	 * bit[34:36]  pad current level
+	 */
 	uint32_t pin: 11;
 	uint32_t pull: 2;
 	uint32_t drive: 1;
@@ -36,13 +39,6 @@ typedef struct {
 
 typedef pinctrl_soc_pin pinctrl_soc_pin_t;
 
-/**
- * @brief Utility macro to initialize each pin.
- *
- * @param node_id Node identifier.
- * @param prop Property name.
- * @param idx Property entry index.
- */
 #define Z_PINCTRL_STATE_PIN_INIT(node_id, prop, idx)                         \
 	{                                                                    \
 		.pin = BEE_GET_PIN(DT_PROP_BY_IDX(node_id, prop, idx)),      \
@@ -56,55 +52,18 @@ typedef pinctrl_soc_pin pinctrl_soc_pin_t;
 		.current_level = DT_PROP_OR(node_id, current_level, 0),                  \
 	},
 
-/**
- * @brief Utility macro to initialize state pins contained in a given property.
- *
- * @param node_id Node identifier.
- * @param prop Property name describing state pins.
- */
 #define Z_PINCTRL_STATE_PINS_INIT(node_id, prop)                                                   \
 	{DT_FOREACH_CHILD_VARGS(DT_PHANDLE(node_id, prop), DT_FOREACH_PROP_ELEM, psels,            \
 				Z_PINCTRL_STATE_PIN_INIT)}
 
-/**
- * @brief Utility macro to obtain pin function.
- *
- * @param pincfg Pin configuration bit field.
- */
 #define BEE_GET_FUN(pincfg) (((pincfg) >> BEE_FUN_POS) & BEE_FUN_MSK)
-
-/**
- * @brief Utility macro to obtain pin drive mode.
- *
- * @param pincfg Pin configuration bit field.
- */
 #define BEE_GET_DIR(pincfg) (((pincfg) >> BEE_DIR_POS) & BEE_DIR_MSK)
-
-/**
- * @brief Utility macro to obtain pin drive mode.
- *
- * @param pincfg Pin configuration bit field.
- */
 #define BEE_GET_DRIVE(pincfg) (((pincfg) >> BEE_DRIVE_POS) & BEE_DRIVE_MSK)
-
-/**
- * @brief Utility macro to obtain pin pull configuration.
- *
- * @param pincfg Pin configuration bit field.
- */
 #define BEE_GET_PULL(pincfg) (((pincfg) >> BEE_PULL_POS) & BEE_PULL_MSK)
-
-/**
- * @brief Utility macro to obtain port and pin combination.
- *
- * @param pincfg Pin configuration bit field.
- */
 #define BEE_GET_PIN(pincfg) (((pincfg) >> BEE_PIN_POS) & BEE_PIN_MSK)
-
-/** @endcond */
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* ZEPHYR_SOC_ARM_REALTEK_RTL_COMMON_PINCTRL_SOC_H_ */
+#endif /* ZEPHYR_SOC_REALTEK_BEE_RTL87X2G_PINCTRL_SOC_H_ */
